requirements.txt
docker-py
deb-pkg-tools
graphviz
prettytable
click
requests
pyyaml
clint
